+++ src/com/intellij/idea/plugin/hybris/common/HybrisConstants.kt
@@ -479,7 +479,7 @@ object HybrisConstants {
     val BACKOFFICE_WEB_INF_CLASSES_2019 = FileUtilRt.toSystemDependentName("/bin/modules/backoffice-framework/backoffice/web/webroot/WEB-INF/classes")
 
     @JvmField
-    val QUERY_STORAGE_FOLDER_PATH = EXCLUDE_IDEA_DIRECTORY + File.separator + "consolestorage"
+    val QUERY_STORAGE_FOLDER_PATH = "consolestorage"
     @JvmField
     val SRC_DIR_NAMES = listOf(SRC_DIRECTORY, GROOVY_SRC_DIRECTORY, KOTLIN_SRC_DIRECTORY, SCALA_SRC_DIRECTORY)
     @JvmField